more review comments, fix tests
